requirements.txt
streamlit==1.31.1
python-dotenv==1.0.0
PyPDF2==3.0.1
requests==2.31.0
python-magic==0.4.27 